# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/dashboard-summary:
    get:
      operationId: getDashboardSummary
      tags: [reports]
      summary: Current month summary - budget, income, expenses, balance, budget progress
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "YYYY-MM format, defaults to current month"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /reports/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [reports]
      summary: Spending breakdown by category for a month
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"
...
  /reports/budget-vs-actual:
    get:
      operationId: getBudgetVsActual
      tags: [reports]
      summary: Budget vs actual spending per category for a month
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: Budget vs actual
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetVsActual"
...
  /reports/triggered-alerts:
    get:
      operationId: getTriggeredAlerts
      tags: [reports]
      summary: Currently triggered budget alerts based on spending vs thresholds
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: Triggered alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TriggeredAlert"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DashboardSummary:
      type: object
      required: [month, totalIncome, totalExpenses, currentBalance, monthlyBudget, budgetSpent, budgetRemaining, budgetPercentUsed]
      properties:
        month:
          type: string
        totalIncome:
          type: number
        totalExpenses:
          type: number
        currentBalance:
          type: number
        monthlyBudget:
          type: ["number", "null"]
        budgetSpent:
          type: number
        budgetRemaining:
          type: ["number", "null"]
        budgetPercentUsed:
          type: ["number", "null"]
    CategorySpending:
      type: object
      required: [categoryId, categoryName, amount, percentage]
      properties:
        categoryId:
          type: number
        categoryName:
          type: string
        amount:
          type: number
        percentage:
          type: number
        color:
          type: ["string", "null"]
...
    BudgetVsActual:
      type: object
      required: [name, budget, actual, remaining, percentUsed, type]
      properties:
        name:
          type: string
        budget:
          type: number
        actual:
          type: number
        remaining:
          type: number
        percentUsed:
          type: number
        type:
          type: string
          enum: [overall, category]
        categoryId:
          type: ["number", "null"]
...
    TriggeredAlert:
      type: object
      required: [alertId, type, threshold, currentPercentage, amount, budgetAmount]
      properties:
        alertId:
          type: number
        type:
          type: string
          enum: [overall, category]
        categoryId:
          type: ["number", "null"]
        categoryName:
          type: ["string", "null"]
        threshold:
          type: number
        currentPercentage:
          type: number
        amount:
          type: number
        budgetAmount:
          type: number
...