# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  currentBalance: number;
  /** @nullable */
  monthlyBudget: number | null;
  budgetSpent: number;
  /** @nullable */
  budgetRemaining: number | null;
  /** @nullable */
  budgetPercentUsed: number | null;
}