# lib/api-zod/src/generated/types/triggeredAlert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { TriggeredAlertType } from './triggeredAlertType';

export interface TriggeredAlert {
  alertId: number;
  type: TriggeredAlertType;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  threshold: number;
  currentPercentage: number;
  amount: number;
  budgetAmount: number;
}